linters dependency cycle fixed
