+++ src/modules/add-new-like.js
@@ -1,34 +1,34 @@
 import postLike from './postLike.js';
 
 const updateLike = async () => {
-    const likeIcons = document.querySelectorAll('.fa-heart');
+  const likeIcons = document.querySelectorAll('.fa-heart');
 
-    function iconClicked(event) {
-        const icon = event.target;
-        const pElement = icon.parentNode;
-        const likeCountElement = pElement.querySelector('.like-count');
+  function iconClicked(event) {
+    const icon = event.target;
+    const pElement = icon.parentNode;
+    const likeCountElement = pElement.querySelector('.like-count');
 
-        let likes = parseInt(icon.dataset.likes);
+    let likes = parseInt(icon.dataset.likes, 10);
 
-        postLike(likes, event);
+    postLike(likes, event);
 
-        icon.classList.toggle('active');
+    icon.classList.toggle('active');
 
-        if (icon.classList.contains('active')) {
-            likes += 1;
-        } else {
-            likes -= 1;
-        }
-
-        icon.dataset.likes = likes;
-        likeCountElement.textContent = likes;
+    if (icon.classList.contains('active')) {
+      likes += 1;
+    } else {
+      likes -= 1;
     }
 
-    likeIcons.forEach((individualIcon) => {
-        individualIcon.addEventListener('click', (event) => {
-            iconClicked(event);
-        });
+    icon.dataset.likes = likes;
+    likeCountElement.textContent = likes;
+  }
+
+  likeIcons.forEach((individualIcon) => {
+    individualIcon.addEventListener('click', (event) => {
+      iconClicked(event);
     });
+  });
 };
 
 export default updateLike;
