+++ src/modules/postLike.js
@@ -1,8 +1,8 @@
-import uniqueId from "..";
+import uniqueId from './uniqueid.js';
 
 const postLike = async (likes, event) => {
-    const newAppId = await uniqueId();
-  const item1Url = `https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/${newAppId}/likes`;  
+  const newAppId = await uniqueId();
+  const item1Url = `https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/${newAppId}/likes`;
   const item1 = event.target.parentElement.className;
   const requestData = {
     method: 'POST',
@@ -11,13 +11,12 @@ const postLike = async (likes, event) => {
     },
     body: JSON.stringify({
       item_id: item1,
-      likes: likes,
+      likes,
     }),
   };
 
-    const response = await fetch(item1Url, requestData);
-    console.log (response);
-
+  const response = await fetch(item1Url, requestData);
+  console.log(response);
 };
 
 export default postLike;
